requirements.txt
django==4.2.10
djangorestframework==3.14.0
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6